LPD-72790 search-experiences-service: remove Registrar and Abstract Model Listener
https://liferay.atlassian.net/browse/LPD-72790
