+++ modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/model/listener/InfoCollectionProviderSXPBlueprintModelListener.java
@@ -5,11 +5,46 @@
 
 package com.liferay.search.experiences.internal.model.listener;
 
+import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.asset.util.AssetHelper;
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.blogs.service.BlogsEntryLocalService;
+import com.liferay.calendar.model.CalendarBooking;
+import com.liferay.calendar.service.CalendarBookingLocalService;
+import com.liferay.document.library.kernel.model.DLFileEntry;
+import com.liferay.document.library.kernel.service.DLAppLocalService;
+import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalService;
+import com.liferay.dynamic.data.mapping.service.DDMStructureService;
 import com.liferay.info.collection.provider.InfoCollectionProvider;
+import com.liferay.journal.model.JournalArticle;
+import com.liferay.journal.service.JournalArticleService;
+import com.liferay.knowledge.base.model.KBArticle;
+import com.liferay.knowledge.base.service.KBArticleLocalService;
+import com.liferay.object.model.ObjectDefinition;
+import com.liferay.object.service.ObjectDefinitionLocalService;
+import com.liferay.object.service.ObjectEntryLocalService;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.model.ModelListener;
+import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.service.GroupService;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.search.asset.AssetSubtypeIdentifier;
+import com.liferay.portal.search.asset.AssetSubtypeIdentifierBuilder;
+import com.liferay.portal.search.searcher.SearchRequestBuilderFactory;
+import com.liferay.portal.search.searcher.Searcher;
+import com.liferay.search.experiences.internal.info.collection.provider.AssetEntrySXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.BlogsEntrySXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.CalendarBookingSXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.FileEntrySXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.JournalArticleSXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.KBArticleSXPBlueprintInfoCollectionProvider;
+import com.liferay.search.experiences.internal.info.collection.provider.ObjectEntrySXPBlueprintInfoCollectionProvider;
 import com.liferay.search.experiences.model.SXPBlueprint;
+import com.liferay.search.experiences.rest.dto.v1_0.Configuration;
+import com.liferay.search.experiences.rest.dto.v1_0.GeneralConfiguration;
 import com.liferay.search.experiences.service.SXPBlueprintLocalService;
 
 import java.util.List;
@@ -18,38 +53,36 @@ import java.util.concurrent.ConcurrentHashMap;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Shuyang Zhou
+ * @author Joshua Cords
  */
-public abstract class InfoCollectionProviderSXPBlueprintModelListener
+@Component(enabled = false, service = ModelListener.class)
+public class InfoCollectionProviderSXPBlueprintModelListener
 	extends BaseModelListener<SXPBlueprint> {
-
-	public InfoCollectionProviderSXPBlueprintModelListener(
-		BundleContext bundleContext, CompanyLocalService companyLocalService,
-		SXPBlueprintLocalService sxpBlueprintLocalService) {
-
-		_bundleContext = bundleContext;
-		_companyLocalService = companyLocalService;
-		_sxpBlueprintLocalService = sxpBlueprintLocalService;
-	}
-
-	@Override
-	public Class<?> getModelClass() {
-		return SXPBlueprint.class;
-	}
 
 	@Override
 	public void onAfterCreate(SXPBlueprint sxpBlueprint) {
+		if (!_isCollectionProvider(sxpBlueprint)) {
+			return;
+		}
+
+		_className = _getClassName(sxpBlueprint);
+
 		ServiceRegistration<?> serviceRegistration = _serviceRegistrations.put(
 			sxpBlueprint.getSXPBlueprintId(),
 			_bundleContext.registerService(
 				InfoCollectionProvider.class,
-				createInfoCollectionProvider(sxpBlueprint),
+				_createInfoCollectionProvider(sxpBlueprint),
 				HashMapDictionaryBuilder.<String, Object>put(
 					"company.id", sxpBlueprint.getCompanyId()
 				).put(
-					"item.class.name", getItemClassName()
+					"item.class.name", _className
 				).build()));
 
 		if (serviceRegistration != null) {
@@ -76,7 +109,10 @@ public abstract class InfoCollectionProviderSXPBlueprintModelListener
 		onAfterCreate(newSXPBlueprint);
 	}
 
-	public void start() {
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_bundleContext = bundleContext;
+
 		_companyLocalService.forEachCompanyId(
 			companyId -> {
 				List<SXPBlueprint> sxpBlueprints =
@@ -88,7 +124,8 @@ public abstract class InfoCollectionProviderSXPBlueprintModelListener
 			});
 	}
 
-	public void stop() {
+	@Deactivate
+	protected void deactivate() {
 		for (ServiceRegistration<?> serviceRegistration :
 				_serviceRegistrations.values()) {
 
@@ -98,15 +135,151 @@ public abstract class InfoCollectionProviderSXPBlueprintModelListener
 		_serviceRegistrations.clear();
 	}
 
-	protected abstract InfoCollectionProvider<?> createInfoCollectionProvider(
-		SXPBlueprint sxpBlueprint);
+	private InfoCollectionProvider<?> _createInfoCollectionProvider(
+		SXPBlueprint sxpBlueprint) {
 
-	protected abstract String getItemClassName();
+		if (_className.equals(BlogsEntry.class.getName())) {
+			return new BlogsEntrySXPBlueprintInfoCollectionProvider(
+				_assetHelper, _blogsEntryLocalService, _searcher,
+				_searchRequestBuilderFactory, sxpBlueprint);
+		}
 
-	private final BundleContext _bundleContext;
-	private final CompanyLocalService _companyLocalService;
+		if (_className.equals(CalendarBooking.class.getName())) {
+			return new CalendarBookingSXPBlueprintInfoCollectionProvider(
+				_assetHelper, _calendarBookingLocalService, _searcher,
+				_searchRequestBuilderFactory, sxpBlueprint);
+		}
+
+		if (_className.equals(DLFileEntry.class.getName())) {
+			return new FileEntrySXPBlueprintInfoCollectionProvider(
+				_assetHelper, _assetSubtypeIdentifierBuilder,
+				_dlAppLocalService, _dlFileEntryTypeLocalService, _groupService,
+				_searcher, _searchRequestBuilderFactory, sxpBlueprint);
+		}
+
+		if (_className.equals(JournalArticle.class.getName())) {
+			return new JournalArticleSXPBlueprintInfoCollectionProvider(
+				_assetHelper, _assetSubtypeIdentifierBuilder,
+				_classNameLocalService, _ddmStructureService, _groupService,
+				_journalArticleService, _searcher, _searchRequestBuilderFactory,
+				sxpBlueprint);
+		}
+
+		if (_className.equals(KBArticle.class.getName())) {
+			return new KBArticleSXPBlueprintInfoCollectionProvider(
+				_assetHelper, _kbArticleLocalService, _searcher,
+				_searchRequestBuilderFactory, sxpBlueprint);
+		}
+
+		if (_className.contains(ObjectDefinition.class.getName())) {
+			return new ObjectEntrySXPBlueprintInfoCollectionProvider(
+				_assetHelper, _className, _objectDefinitionLocalService,
+				_objectEntryLocalService, _searcher,
+				_searchRequestBuilderFactory, sxpBlueprint);
+		}
+
+		return new AssetEntrySXPBlueprintInfoCollectionProvider(
+			_assetHelper, _searcher, _searchRequestBuilderFactory,
+			sxpBlueprint);
+	}
+
+	private String _getClassName(SXPBlueprint sxpBlueprint) {
+		Configuration configuration = Configuration.unsafeToDTO(
+			sxpBlueprint.getConfigurationJSON());
+
+		GeneralConfiguration generalConfiguration =
+			configuration.getGeneralConfiguration();
+
+		if (generalConfiguration == null) {
+			return AssetEntry.class.getName();
+		}
+
+		String collectionProviderType =
+			generalConfiguration.getCollectionProviderType();
+
+		if (Validator.isNull(collectionProviderType)) {
+			return AssetEntry.class.getName();
+		}
+
+		AssetSubtypeIdentifier assetSubtypeIdentifier =
+			_assetSubtypeIdentifierBuilder.searchableAssetType(
+				collectionProviderType
+			).build();
+
+		return assetSubtypeIdentifier.getClassName();
+	}
+
+	private boolean _isCollectionProvider(SXPBlueprint sxpBlueprint) {
+		Configuration configuration = Configuration.unsafeToDTO(
+			sxpBlueprint.getConfigurationJSON());
+
+		GeneralConfiguration generalConfiguration =
+			configuration.getGeneralConfiguration();
+
+		if (generalConfiguration == null) {
+			return false;
+		}
+
+		return GetterUtil.getBoolean(
+			generalConfiguration.getCollectionProvider());
+	}
+
+	@Reference
+	private AssetHelper _assetHelper;
+
+	@Reference
+	private AssetSubtypeIdentifierBuilder _assetSubtypeIdentifierBuilder;
+
+	@Reference
+	private BlogsEntryLocalService _blogsEntryLocalService;
+
+	private BundleContext _bundleContext;
+
+	@Reference
+	private CalendarBookingLocalService _calendarBookingLocalService;
+
+	private String _className = AssetEntry.class.getName();
+
+	@Reference
+	private ClassNameLocalService _classNameLocalService;
+
+	@Reference
+	private CompanyLocalService _companyLocalService;
+
+	@Reference
+	private DDMStructureService _ddmStructureService;
+
+	@Reference
+	private DLAppLocalService _dlAppLocalService;
+
+	@Reference
+	private DLFileEntryTypeLocalService _dlFileEntryTypeLocalService;
+
+	@Reference
+	private GroupService _groupService;
+
+	@Reference
+	private JournalArticleService _journalArticleService;
+
+	@Reference
+	private KBArticleLocalService _kbArticleLocalService;
+
+	@Reference
+	private ObjectDefinitionLocalService _objectDefinitionLocalService;
+
+	@Reference
+	private ObjectEntryLocalService _objectEntryLocalService;
+
+	@Reference
+	private Searcher _searcher;
+
+	@Reference
+	private SearchRequestBuilderFactory _searchRequestBuilderFactory;
+
 	private final Map<Long, ServiceRegistration<?>> _serviceRegistrations =
 		new ConcurrentHashMap<>();
-	private final SXPBlueprintLocalService _sxpBlueprintLocalService;
+
+	@Reference
+	private SXPBlueprintLocalService _sxpBlueprintLocalService;
 
 }
